addition of natural numbers

+++ 0x02-functions_nested_loops/101-natural.c
@@ -10,7 +10,7 @@
 
 int main(void)
 {
-long int x, sum;
+int x, sum;
 sum = 0;
 
 for (x = 0 ; x <= 1024 ; x++)
